refractor: organize routes
